backend/requirements.txt
# 1688 商品爬虫依赖
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
sqlite-utils>=3.36
tqdm>=4.66.0
fake-useragent>=1.4.0
pyyaml>=6.0

# US_02 反爬策略依赖
# playwright>=1.40.0  # 用于 Stealth 模式（可选，二期使用）
# playwright-stealth>=1.0.0  # Stealth 插件（可选，二期使用）
